NH0098

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo fechas de operacion.

Usage

NH0098

Format

Un data frame con estas 35 variables:

id

ID de la estacion, datos categoricos.

fecha

Fecha de la medicion, datos temporales.

temperatura_abrigo_150cm

Temperatura medida en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_maxima

Temperatura maxima en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_minima

Temperatura minima en abrigo a 150 cm, datos numericos.

temperatura_intemperie_5cm_minima

Temperatura minima medida en intemperie a 5 cm del suelo, valores NA.

temperatura_intemperie_50cm_minima

Temperatura minima medida en intemperie a 50 cm del suelo, valores NA.

temperatura_suelo_5cm_media

Temperatura media medida en el suelo a 5 cm de profundidad, valores NA.

temperatura_suelo_10cm_media

Temperatura media medida en el suelo a 10 cm de profundidad, valores NA.

temperatura_inte_5cm

Temperatura medida a 5 cm en intemperie, valores NA.

temperatura_intemperie_150cm_minima

Temperatura minima en intemperie a 150 cm, valores NA.

humedad_suelo

Humedad del suelo, valores NA.

precipitacion_pluviometrica

Precipitacion pluviometrica en mm, datos numericos.

granizo

Indica presencia de granizo, valores NA.

nieve

Indica presencia de nieve, valores NA.

heliofania_efectiva

Heliofania efectiva, valores NA.

heliofania_relativa

Heliofania relativa, valores NA.

tesion_vapor_media

Tensión de vapor media, valores NA.

humedad_media

Humedad relativa media, valores NA.

humedad_media_8_14_20

Humedad relativa media registrada a las 8, 14 y 20 horas, valores NA.

rocio_medio

Punto de rocio medio, valores NA.

duracion_follaje_mojado

Duracion del follaje mojado, valores NA.

velocidad_viento_200cm_media

Velocidad media del viento a 200 cm de altura, valores NA.

direccion_viento_200cm

Direccion del viento a 200 cm de altura, valores NA.

velocidad_viento_1000cm_media

Velocidad media del viento a 1000 cm de altura, valores NA.

direccion_viento_1000cm

Direccion del viento a 1000 cm de altura, valores NA.

velocidad_viento_maxima

Velocidad maxima del viento registrada, valores NA.

presion_media

Presion atmosferica media, valores NA.

radiacion_global

Radiacion solar global medida en la estacion, valores NA.

radiacion_neta

Radiacion neta medida en la estacion, valores NA.

evaporacion_tanque

Evaporacion medida en tanque, valores NA.

evapotranspiracion_potencial

Evapotranspiracion potencial, valores NA.

profundidad_napa

Profundidad de la napa freatica, valores NA.

horas_frio

Cantidad de horas de frio registradas, valores numéricos.

unidad_frio

Unidad de medida para las horas de frio, valores NA.

Source

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.